ROC-1183 - Changed int type assertion to uint8 for node_streamrequest.go.

+++ node_streamrequest.go
@@ -88,7 +88,7 @@ func (sr *nodeStreamRequest) onEventFrame(evt *EventFrame) {
 	// message must be heartbeat and sender must be an ardupilot device
 
 	if msg, ok := evt.Message().(*DynamicMessage); ok {
-		if msg.GetId() != 0 || msg.Fields["Autopilot"].(int) != 3 {
+		if msg.GetId() != 0 || msg.Fields["Autopilot"].(uint8) != 3 {
 			return
 		}
 	} else {
